Move mouse release regression to Swift Testing
